+++ bd/03-TABLA_INFORME.sql
@@ -1,10 +1,10 @@
 -- CREACION DE LA TABLA DE ESPECIFICACIONES DE CADA USUARIO
 CREATE TABLE T_INFORME(
-	ID_INFORME INT AUTO_INCREMENT PRIMARY KEY COMMENT 'Identificador de la tabla de informe del perfil deportista del usuario.',
-	FK_OBJETIVO INT NOT NULL COMMENT 'Identificador FK para la tabla objetivo',
-	N_EDAD INT NOT NULL COMMENT 'Edad del usuario que se somete al cambio físico',
+	ID_INFORME BIGINT AUTO_INCREMENT PRIMARY KEY COMMENT 'Identificador de la tabla de informe del perfil deportista del usuario.',
+	FK_OBJETIVO BIGINT NOT NULL COMMENT 'Identificador FK para la tabla objetivo',
+	N_EDAD BIGINT NOT NULL COMMENT 'Edad del usuario que se somete al cambio físico',
 	D_SEXO ENUM('Masculino', 'Femenino') NOT NULL COMMENT 'Define el sexo del usuario (`Masculino`, `Femenino`)',
-	N_ALTURA INT NOT NULL COMMENT 'Altura del usuario, en centímetros',
+	N_ALTURA BIGINT NOT NULL COMMENT 'Altura del usuario, en centímetros',
 	N_PESO DECIMAL(5,2) NOT NULL COMMENT 'Peso del usuario, con hasta 2 decimales',
 	N_PORCENTAJE_GRASO DECIMAL(4,2) COMMENT 'Porcentaje graso corporal del usuario. Opcional',
 	N_PORCENTAJE_MUSCULO DECIMAL(4,2) COMMENT 'Porcentaje de masa muscular del usuario. Opcional',
